Added a scrapy code program
This program crawls subcategory (and its subcategories) from the domain http://dmoztools.net. The dmoztools.net is the directory of webpages for different categories.
At this stage, it will crawl dmoztools.net/computer/freeware and its subcategories for (web)sites listed on each of the pages. This program can run until it has crawled all the pages unless the spider is stopped by Ctrl+C. As this can cause too much traffic, further versions of this program will have extensions/spider settings to stop the spider when it reaches a certain limit.